+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Improve the error message of [`queryMaybeOne`](README.md#queryMaybeOne) in cases where
+  the query returns an unexpected amount of rows.
+- Improve documentation
+
 ## [0.9.0] - 2021-01-26
 
 ### Added